+++ qbraid/runtime/native/provider.py
@@ -25,8 +25,9 @@ from qbraid.programs import QPROGRAM_REGISTRY, ProgramSpec
 from qbraid.programs.circuits.qasm import has_measurements
 from qbraid.programs.typer import Qasm2StringType, Qasm3StringType
 from qbraid.runtime._display import display_jobs_from_data
-from qbraid.runtime.enums import ExperimentType, NoiseModel, NoiseModelWrapper
+from qbraid.runtime.enums import ExperimentType
 from qbraid.runtime.exceptions import ResourceNotFoundError
+from qbraid.runtime.noise import NoiseModelSet
 from qbraid.runtime.profile import TargetProfile
 from qbraid.runtime.provider import QuantumProvider
 from qbraid.runtime.schemas.device import DeviceData
@@ -121,32 +122,14 @@ class QbraidProvider(QuantumProvider):
         lambdas = get_program_spec_lambdas(run_package, device_id)
         return ProgramSpec(program_type, alias=run_package, **lambdas) if program_type else None
 
-    @staticmethod
-    def _process_noise_models(noise_models: Optional[list[str]]) -> Optional[set[NoiseModel]]:
-        """Process and return a list of noise model enum objects."""
-        if not noise_models:
-            return None
-
-        noise_model_enums = []
-
-        for nm in noise_models:
-            nm = nm.lower()
-            try:
-                nm_enum = NoiseModel(nm)
-            except ValueError:
-                NoiseModelWrapper.set_other_value(nm)
-                nm_enum = NoiseModelWrapper.get_noise_model(NoiseModel.Other)
-
-            noise_model_enums.append(nm_enum)
-
-        return set(noise_model_enums)
-
     def _build_runtime_profile(self, device_data: dict[str, Any]) -> TargetProfile:
         """Builds a runtime profile from qBraid device data."""
         model = DeviceData(**device_data)
         simulator = str(model.device_type).upper() == "SIMULATOR"
         program_spec = self._get_program_spec(model.run_package, model.device_id)
-        noise_models = self._process_noise_models(model.noise_models)
+        noise_models = (
+            NoiseModelSet.from_iterable(model.noise_models) if model.noise_models else None
+        )
         device_exp_type = "gate_model" if model.paradigm == "gate-based" else model.paradigm.lower()
         experiment_type = ExperimentType(device_exp_type)
         return TargetProfile(
@@ -158,7 +141,7 @@ class QbraidProvider(QuantumProvider):
             provider_name=model.provider,
             noise_models=noise_models,
             name=model.name,
-            pricing=model.pricing.model_dump(),
+            pricing=model.pricing,
         )
 
     @cached_method(ttl=120)